+++ var/httpd/htdocs/js/Core.Config.js
@@ -151,25 +151,7 @@ Core.Config = (function (TargetNS) {
             return ($.browser.msie && $.browser.version === '6.0');
         },
         'Microsoft Internet Explorer 7': function () {
-            var detected = ($.browser.msie && $.browser.version === '7.0');
-
-            // IE8 in Compatibility Mode will claim to be IE7.
-            // See also http://msdn.microsoft.com/en-us/library/ms537503%28v=VS.85%29.aspx
-            if (detected && navigator && navigator.userAgent && navigator.userAgent.match(/Trident\/4.0/)) {
-                alert('Please turn off Compatibility Mode in Internet Explorer.');
-            }
-
-            // IE9 in Compatibility Mode will claim to be IE7.
-            if (detected && navigator && navigator.userAgent && navigator.userAgent.match(/Trident\/5.0/)) {
-                alert('Please turn off Compatibility Mode in Internet Explorer.');
-            }
-
-            // IE10 in Compatibility Mode will claim to be IE7.
-            if (detected && navigator && navigator.userAgent && navigator.userAgent.match(/Trident\/6.0/)) {
-                alert('Please turn off Compatibility Mode in Internet Explorer.');
-            }
-
-            return detected;
+            return ($.browser.msie && $.browser.version === '7.0');
         },
         'Konqueror (without WebKit engine)': function () {
             return ($.browser.webkit && navigator.vendor === 'KDE');
@@ -202,25 +184,7 @@ Core.Config = (function (TargetNS) {
             return ($.browser.msie && $.browser.version === '6.0');
         },
         'Microsoft Internet Explorer 7': function () {
-            var detected = ($.browser.msie && $.browser.version === '7.0');
-
-            // IE8 in Compatibility Mode will claim to be IE7.
-            // See also http://msdn.microsoft.com/en-us/library/ms537503%28v=VS.85%29.aspx
-            if (detected && navigator && navigator.userAgent && navigator.userAgent.match(/Trident\/4.0/)) {
-                alert('Please turn off Compatibility Mode in Internet Explorer.');
-            }
-
-            // IE9 in Compatibility Mode will claim to be IE7.
-            if (detected && navigator && navigator.userAgent && navigator.userAgent.match(/Trident\/5.0/)) {
-                alert('Please turn off Compatibility Mode in Internet Explorer.');
-            }
-
-            // IE10 in Compatibility Mode will claim to be IE7.
-            if (detected && navigator && navigator.userAgent && navigator.userAgent.match(/Trident\/6.0/)) {
-                alert('Please turn off Compatibility Mode in Internet Explorer.');
-            }
-
-            return detected;
+            return ($.browser.msie && $.browser.version === '7.0');
         },
         'Konqueror (without WebKit engine)': function () {
             return ($.browser.webkit && navigator.vendor === 'KDE');